EntregaProyecto

+++ src/main/java/com/sofka/qa/TerminalTransporteApp/Models/Pasajero.java
@@ -6,18 +6,15 @@ public abstract class Pasajero {
     private String nombre;
     private String apellido;
 
-    public Long getId() {
-        return id;
+    public Integer getId() {
+        return Math.toIntExact(id);
     }
-
     public void setId(Long id) {
         this.id = id;
     }
-
     public String getNombre() {
         return nombre;
     }
-
     public void setNombre(String nombre) {
         this.nombre = nombre;
     }
